+++ Tag.WPF/Tagging/TaggingViewModel.cs
@@ -264,11 +264,14 @@ namespace Tag.WPF
                             filename = dir + @"\" + filename + ext;
                             try
                             {
-                                if (System.IO.File.Exists(filename))
+                                if (Path.GetFullPath(filename) != Path.GetFullPath(value.TagInfo.Path))
                                 {
-                                    System.IO.File.Delete(filename);
+                                    if (System.IO.File.Exists(filename))
+                                    {
+                                        System.IO.File.Delete(filename);
+                                    }
+                                    System.IO.File.Move(value.TagInfo.Path, filename);
                                 }
-                                System.IO.File.Move(value.TagInfo.Path, filename);
                             }
                             catch { }
                             t.Add(filename);
